Move streaming utils to parent
